Module pancake::storage::serde

Serialization format

Following pseudocode depicts the byte representation on disk. The actual struct definitions do not exit.

This file format is applicable for both the commit log and ss tables.

struct File {
    k0: Item,
    v0: Item,
    k1: Item,
    v1: Item,
    ...
    // There are no separators in between Items and nothing to indicate
    // whether an Item is a key or a value.
}

struct Item {
    // Encodes the on-disk byte size of `datum_type` and `datum`.
    // Span size is placed first, so that a reader that is uninterested
    // in deserializing datum can skip over it.
    span_size: usize,

    // Identifies how `datum` should be deserialized.
    // The choice of u8 is arbitrary. In case we need to deprecate
    // supported datum_types over time, this allows us
    // (pow(2, 8) - count_of_active_datum_types) deprecations, before
    // rolling over to zero.
    datum_type: u8,

    // Datum may be empty. Examples: tombstone, empty string, empty
    // map, custom serialization.
    datum: [u8; variable_length],
}

A tuple is a type of datum that can nest other data, including other tuple-typed data. For a tuple, the datum byte sequence encodes the following structure:

struct TupleDatum {
    length_of_tuple: usize,
    member_0: Item,
    member_1: Item,
    ...
    member_n-1: Item,
}

Structs

KeyValueIterator

Enums

Datum

DatumType

We manually map enum members to data_type integers because:

OptDatum

ReadItem

SkipItem

Traits

Serializable

Functions

read_item

serialize_kv

skip_item

write_item

@return Total count of bytes that are written to file.